+++ Plugins/PLCompositing/src/Shaders/General/SRPDirectionalLightingShaders_GLSL.h
@@ -23,53 +23,53 @@
 // GLSL (OpenGL 2.1 ("#version 120") and OpenGL ES 2.0 ("#version 100")) vertex shader source code, "#version" is added by "PLRenderer::ProgramGenerator"
 static const PLCore::String sDirectionalLighting_GLSL_VS = "\
 // In attributes\n\
-attribute highp vec4 VertexPosition;			// Object space vertex position input\n\
+attribute mediump vec4 VertexPosition;			// Object space vertex position input\n\
 #ifdef VS_TEXCOORD0\n\
-	attribute highp vec2 VertexTexCoord0;		// Vertex texture coordinate 0 input\n\
+	attribute mediump vec2 VertexTexCoord0;		// Vertex texture coordinate 0 input\n\
 #endif\n\
 #ifdef VS_TEXCOORD1\n\
-	attribute highp vec2 VertexTexCoord1;		// Vertex texture coordinate 1 input\n\
+	attribute mediump vec2 VertexTexCoord1;		// Vertex texture coordinate 1 input\n\
 #endif\n\
 #ifdef VS_NORMAL\n\
-	attribute highp vec3 VertexNormal;			// Object space vertex normal input\n\
+	attribute mediump vec3 VertexNormal;		// Object space vertex normal input\n\
 	#ifdef VS_TANGENT_BINORMAL\n\
-		attribute highp vec3 VertexTangent;		// Object space vertex tangent input\n\
-		attribute highp vec3 VertexBinormal;	// Object space vertex binormal input\n\
+		attribute mediump vec3 VertexTangent;	// Object space vertex tangent input\n\
+		attribute mediump vec3 VertexBinormal;	// Object space vertex binormal input\n\
 	#endif\n\
 #endif\n\
 \n\
 // Out attributes\n\
 #if defined(VS_TEXCOORD0) && defined(VS_TEXCOORD1)\n\
-	varying highp vec4 VertexTexCoordVS;		// Vertex texture coordinate 0 (xy) and 1 (zw) output\n\
+	varying mediump vec4 VertexTexCoordVS;		// Vertex texture coordinate 0 (xy) and 1 (zw) output\n\
 #else\n\
 	#if defined(VS_TEXCOORD0) || defined(VS_TEXCOORD1)\n\
-		varying highp vec2 VertexTexCoordVS;	// Vertex texture coordinate 0 or 1 output\n\
+		varying mediump vec2 VertexTexCoordVS;	// Vertex texture coordinate 0 or 1 output\n\
 	#endif\n\
 #endif\n\
 #ifdef VS_VIEWSPACEPOSITION\n\
-	varying highp vec3 VertexPositionVS;		// View space vertex position output\n\
+	varying mediump vec3 VertexPositionVS;		// View space vertex position output\n\
 #endif\n\
 #ifdef VS_NORMAL\n\
-	varying highp vec3 VertexNormalVS;			// View space vertex normal output\n\
+	varying mediump vec3 VertexNormalVS;		// View space vertex normal output\n\
 	#ifdef VS_TANGENT_BINORMAL\n\
-		varying highp vec3 VertexTangentVS;		// View space vertex tangent output\n\
-		varying highp vec3 VertexBinormalVS;	// View space vertex binormal output\n\
+		varying mediump vec3 VertexTangentVS;	// View space vertex tangent output\n\
+		varying mediump vec3 VertexBinormalVS;	// View space vertex binormal output\n\
 		#ifdef VS_PARALLAXMAPPING\n\
-			varying highp vec3 EyeVecVS;		// Tangent space vector pointing from the pixel to the eye point\n\
+			varying mediump vec3 EyeVecVS;		// Tangent space vector pointing from the pixel to the eye point\n\
 		#endif\n\
 	#endif\n\
 #endif\n\
 \n\
 // Uniforms\n\
 #if defined(VS_NORMAL) && defined(VS_TWOSIDEDLIGHTING)\n\
-	uniform highp float NormalScale;					// Normal scale (negative to flip normals)\n\
+	uniform mediump float NormalScale;					// Normal scale (negative to flip normals)\n\
 #endif\n\
 #if defined(VS_VIEWSPACEPOSITION) || defined(VS_NORMAL)\n\
-	uniform highp mat4 ObjectSpaceToViewSpaceMatrix;	// Object space to view space matrix\n\
+	uniform mediump mat4 ObjectSpaceToViewSpaceMatrix;	// Object space to view space matrix\n\
 #endif\n\
-uniform highp mat4 ObjectSpaceToClipSpaceMatrix;		// Object space to clip space matrix\n\
+uniform mediump mat4 ObjectSpaceToClipSpaceMatrix;		// Object space to clip space matrix\n\
 #if defined(VS_NORMAL) && defined(VS_TANGENT_BINORMAL) && defined(VS_PARALLAXMAPPING)\n\
-	uniform highp vec3 EyePos;							// Object space eye position\n\
+	uniform mediump vec3 EyePos;						// Object space eye position\n\
 #endif\n\
 \n\
 // Programs\n\
@@ -81,7 +81,7 @@ void main()\n\
 #ifdef VS_TEXCOORD0\n\
 	#ifdef VS_TEXCOORD1\n\
 		// Pass through the vertex texture coordinate 0 and 1 together\n\
-		VertexTexCoordVS = highp vec4(VertexTexCoord0, VertexTexCoord1);\n\
+		VertexTexCoordVS = mediump vec4(VertexTexCoord0, VertexTexCoord1);\n\
 	#else\n\
 		// Pass through the vertex texture coordinate 0\n\
 		VertexTexCoordVS = VertexTexCoord0;\n\
@@ -95,7 +95,7 @@ void main()\n\
 \n\
 #ifdef VS_VIEWSPACEPOSITION\n\
 	// Transform vertex position to view space\n\
-	highp vec4 viewSpacePosition = ObjectSpaceToViewSpaceMatrix*VertexPosition;\n\
+	mediump vec4 viewSpacePosition = ObjectSpaceToViewSpaceMatrix*VertexPosition;\n\
 	VertexPositionVS = viewSpacePosition.xyz;\n\
 #endif\n\
 \n\
@@ -126,22 +126,22 @@ void main()\n\
 static const PLCore::String sDirectionalLighting_GLSL_FS = "\
 // Attributes\n\
 #if defined(FS_TEXCOORD0) && defined(FS_TEXCOORD1)\n\
-	varying highp vec4 VertexTexCoordVS;		// Vertex texture coordinate 0 (xy) and 1 (zw) from vertex shader\n\
+	varying mediump vec4 VertexTexCoordVS;		// Vertex texture coordinate 0 (xy) and 1 (zw) from vertex shader\n\
 #else\n\
 	#if defined(FS_TEXCOORD0) || defined(FS_TEXCOORD1)\n\
-		varying highp vec2 VertexTexCoordVS;	// Vertex texture coordinate 0 or 1 from vertex shader\n\
+		varying mediump vec2 VertexTexCoordVS;	// Vertex texture coordinate 0 or 1 from vertex shader\n\
 	#endif\n\
 #endif\n\
 #if defined(FS_REFLECTION) || defined(FS_LIGHTING) || defined(FS_DOF)\n\
-	varying highp vec3 VertexPositionVS;		// View space vertex position from vertex shader\n\
+	varying mediump vec3 VertexPositionVS;		// View space vertex position from vertex shader\n\
 #endif\n\
 #ifdef FS_NORMAL\n\
-	varying highp vec3 VertexNormalVS;			// View space vertex normal from vertex shader\n\
+	varying mediump vec3 VertexNormalVS;		// View space vertex normal from vertex shader\n\
 	#ifdef FS_TANGENT_BINORMAL\n\
-		varying highp vec3 VertexTangentVS;		// View space vertex tangent from vertex shader\n\
-		varying highp vec3 VertexBinormalVS;	// View space vertex binormal from vertex shader\n\
+		varying mediump vec3 VertexTangentVS;	// View space vertex tangent from vertex shader\n\
+		varying mediump vec3 VertexBinormalVS;	// View space vertex binormal from vertex shader\n\
 		#ifdef FS_PARALLAXMAPPING\n\
-			varying highp vec3 EyeVecVS;		// Tangent space vector pointing from the pixel to the eye point from vertex shader\n\
+			varying mediump vec3 EyeVecVS;		// Tangent space vector pointing from the pixel to the eye point from vertex shader\n\
 		#endif\n\
 	#endif\n\
 #endif\n\
@@ -156,7 +156,7 @@ uniform lowp vec4 DiffuseColor;	// Alpha stores the opacity\n\
 	#endif\n\
 #endif\n\
 #ifdef FS_DIFFUSERAMPMAP\n\
-     uniform highp sampler1D DiffuseRampMap;\n\
+     uniform mediump sampler1D DiffuseRampMap;\n\
 #endif\n\
 #ifdef FS_AMBIENTOCCLUSIONMAP\n\
 	uniform lowp sampler2D AmbientOcclusionMap;\n\
@@ -177,14 +177,14 @@ uniform lowp vec4 DiffuseColor;	// Alpha stores the opacity\n\
 		uniform lowp sampler2D ReflectivityMap;\n\
 	#endif\n\
 	#ifdef FS_FRESNELREFLECTION\n\
-		uniform highp vec2 FresnelConstants; // x = R0 [0..1] and y = Power, always >0\n\
+		uniform mediump vec2 FresnelConstants; // x = R0 [0..1] and y = Power, always >0\n\
 	#endif\n\
 	#ifdef FS_2DREFLECTIONMAP\n\
 		uniform lowp  sampler2D ReflectionMap;\n\
-		uniform highp mat3      ViewSpaceToWorldSpace;		// View space to world space transform matrix\n\
+		uniform mediump mat3      ViewSpaceToWorldSpace;	// View space to world space transform matrix\n\
 	#elif defined(FS_CUBEREFLECTIONMAP)\n\
 		uniform lowp  samplerCube ReflectionMap;\n\
-		uniform highp mat3        ViewSpaceToWorldSpace;	// View space to world space transform matrix\n\
+		uniform mediump mat3        ViewSpaceToWorldSpace;	// View space to world space transform matrix\n\
 	#endif\n\
 #endif\n\
 #ifdef FS_NORMALMAP\n\
@@ -201,8 +201,8 @@ uniform lowp vec4 DiffuseColor;	// Alpha stores the opacity\n\
 	uniform mediump vec2      ParallaxScaleBias;\n\
 #endif\n\
 #ifdef FS_LIGHTING\n\
-	uniform highp vec3 LightDirection;	// View space light direction\n\
-	uniform lowp  vec3 LightColor;\n\
+	uniform mediump vec3 LightDirection;	// View space light direction\n\
+	uniform lowp    vec3 LightColor;\n\
 	#ifdef FS_SPECULAR\n\
 		uniform lowp vec3  SpecularColor;\n\
 		uniform lowp float SpecularExponent;\n\
@@ -210,11 +210,11 @@ uniform lowp vec4 DiffuseColor;	// Alpha stores the opacity\n\
 			uniform lowp sampler2D SpecularMap;\n\
 		#endif\n\
 		#ifdef FS_SPECULARRAMPMAP\n\
-			uniform highp sampler1D SpecularRampMap;\n\
+			uniform mediump sampler1D SpecularRampMap;\n\
 		#endif\n\
 	#endif\n\
 	#ifdef FS_EDGERAMPMAP\n\
-		uniform highp sampler1D EdgeRampMap;\n\
+		uniform mediump sampler1D EdgeRampMap;\n\
 	#endif\n\
 #endif\n\
 #ifdef FS_GLOW\n\
@@ -227,35 +227,35 @@ uniform lowp vec4 DiffuseColor;	// Alpha stores the opacity\n\
 	// DOFParams coefficients:\n\
 	// x = near blur depth; y = focal plane depth; z = far blur depth\n\
 	// w = blurriness cutoff constant for objects behind the focal plane\n\
-	uniform highp vec4 DOFParams;\n\
+	uniform mediump vec4 DOFParams;\n\
 #endif\n\
 \n\
 // Programs\n\
 #ifdef FS_FRESNELREFLECTION\n\
-	highp float fresnel(highp vec3 light, highp vec3 normal, highp vec2 constants)\n\
+	mediump float fresnel(mediump vec3 light, mediump vec3 normal, mediump vec2 constants)\n\
 	{\n\
 		// Light and normal are assumed to be normalized\n\
 		// constants.x = R0 [0..1]\n\
 		// constants.y = Power, always >0\n\
-		highp float cosAngle = clamp(1.0 - dot(light, normal), 0.0, 1.0); // We REALLY need to clamp in here or pow may hurt us when using negative numbers!\n\
+		mediump float cosAngle = clamp(1.0 - dot(light, normal), 0.0, 1.0); // We REALLY need to clamp in here or pow may hurt us when using negative numbers!\n\
 		return constants.x + (1.0 - constants.x) * pow(cosAngle, constants.y);\n\
 	}\n\
 #endif\n\
 \n\
 #ifdef FS_LIGHTING\n\
 	// Performs the Blinn-Phong lighting calculation\n\
-	lowp vec3 BlinnPhong(highp vec3 lightVector, lowp vec3 lightColor, highp vec3 viewVector, highp vec3 normalVector, lowp vec3 diffuseColor, lowp vec3 specularColor, lowp float specularExponent)\n\
+	lowp vec3 BlinnPhong(mediump vec3 lightVector, lowp vec3 lightColor, mediump vec3 viewVector, mediump vec3 normalVector, lowp vec3 diffuseColor, lowp vec3 specularColor, lowp float specularExponent)\n\
 	{\n\
 		// [TODO] There seem to be invalid normal vectors here (NAN) - IEEE standard: NaN != NaN - I don't use isnan so I can use lower shader versions\n\
 		if (normalVector.x != normalVector.x || normalVector.y != normalVector.y || normalVector.z != normalVector.z)\n\
-			normalVector = highp vec3(0.0, 0.0, 1.0);\n\
+			normalVector = mediump vec3(0.0, 0.0, 1.0);\n\
 		if (lightVector.x != lightVector.x || lightVector.y != lightVector.y || lightVector.z != lightVector.z)\n\
-			lightVector = highp vec3(0.0, 0.0, 1.0);\n\
+			lightVector = mediump vec3(0.0, 0.0, 1.0);\n\
 		if (viewVector.x != viewVector.x || viewVector.y != viewVector.y || viewVector.z != viewVector.z)\n\
-			viewVector = highp vec3(0.0, 0.0, 1.0);\n\
+			viewVector = mediump vec3(0.0, 0.0, 1.0);\n\
 \n\
 		// Diffuse term\n\
-		highp float diffuse = clamp(dot(lightVector, normalVector), 0.0, 1.0);\n\
+		mediump float diffuse = clamp(dot(lightVector, normalVector), 0.0, 1.0);\n\
 		#ifdef FS_DIFFUSERAMPMAP\n\
 			diffuse *= texture1D(DiffuseRampMap, diffuse).x;\n\
 		#endif\n\
@@ -264,7 +264,7 @@ uniform lowp vec4 DiffuseColor;	// Alpha stores the opacity\n\
 		// Specular term\n\
 		#ifdef FS_SPECULAR\n\
 			// Calculate the half vector between the light vector and the view vector. This is cheaper then calculating the actual reflective vector.\n\
-			highp vec3 halfVector = normalize(lightVector + viewVector);\n\
+			mediump vec3 halfVector = normalize(lightVector + viewVector);\n\
 \n\
 			// Ensure that the specular exponent is never ever <=0, else NANs may be produced by pow!\n\
 			#define FLT_MIN 1.175494351e-38 // Minimum positive value\n\
@@ -273,13 +273,13 @@ uniform lowp vec4 DiffuseColor;	// Alpha stores the opacity\n\
 			#undef FLT_MIN\n\
 \n\
 			// Specular term\n\
-			highp float specular = pow(clamp(dot(halfVector, normalVector), 0.0, 1.0), specularExponent);\n\
+			mediump float specular = pow(clamp(dot(halfVector, normalVector), 0.0, 1.0), specularExponent);\n\
 			lowp vec3 specularLighting = specular*specularColor*lightColor;\n\
 			#ifdef FS_SPECULARRAMPMAP\n\
 				specularLighting *= texture1D(SpecularRampMap, specular).x;\n\
 			#endif\n\
 		#else\n\
-			#define specularLighting highp vec3(0.0, 0.0, 0.0)\n\
+			#define specularLighting mediump vec3(0.0, 0.0, 0.0)\n\
 		#endif\n\
 \n\
 		// Edge detection\n\
@@ -298,11 +298,11 @@ void main()\n\
 {\n\
 	// Define were the texture coordinates come from\n\
 	#if defined(FS_TEXCOORD0) && defined(FS_TEXCOORD1)\n\
-		highp vec2 TexCoord0 = VertexTexCoordVS.xy;\n\
+		mediump vec2 TexCoord0 = VertexTexCoordVS.xy;\n\
 		#define TexCoord1 VertexTexCoordVS.zw\n\
 	#else\n\
 		#ifdef FS_TEXCOORD0\n\
-			highp vec2 TexCoord0 = VertexTexCoordVS;\n\
+			mediump vec2 TexCoord0 = VertexTexCoordVS;\n\
 		#endif\n\
 		#ifdef FS_TEXCOORD1\n\
 			#define TexCoord1 VertexTexCoordVS\n\
@@ -314,14 +314,14 @@ void main()\n\
 		// where the ray would hit the surface. The height map indicates how far to offset.\n\
 		// The approximation assumes that all heights in the height-map is equal, which of\n\
 		// course won't be the case, but if the bump map is fairly smooth it works well enough.\n\
-		highp vec3  eyeVec = normalize(EyeVecVS);\n\
-		lowp  float scale  = ParallaxScaleBias.x;\n\
-		lowp  float bias   = ParallaxScaleBias.y;\n\
-		lowp  float height = texture2D(HeightMap, TexCoord0).r;\n\
+		mediump vec3  eyeVec = normalize(EyeVecVS);\n\
+		lowp    float scale  = ParallaxScaleBias.x;\n\
+		lowp    float bias   = ParallaxScaleBias.y;\n\
+		lowp    float height = texture2D(HeightMap, TexCoord0).r;\n\
 \n\
 		// First offset pass with offset limiting (no division through eyeVec.z)\n\
 		lowp float offset = height*scale + bias;\n\
-		highp vec2 textureCoordinate = TexCoord0 + offset*eyeVec.xy;\n\
+		mediump vec2 textureCoordinate = TexCoord0 + offset*eyeVec.xy;\n\
 \n\
 		// For better quality: Refine the parallax by making another lookup at where we ended\n\
 		// up in the first parallax computation, then averaging the results.\n\
@@ -341,22 +341,22 @@ void main()\n\
 	#ifdef FS_LIGHTMAP\n\
 		lowp vec3 lightMapColor = texture2D(LightMap, TexCoord1).rgb;\n\
 		#ifdef FS_GAMMACORRECTION\n\
-			lightMapColor = pow(lightMapColor, highp vec3(2.2, 2.2, 2.2));	// Perform sRGB to linear space conversion (gamma correction)\n\
+			lightMapColor = pow(lightMapColor, mediump vec3(2.2, 2.2, 2.2));	// Perform sRGB to linear space conversion (gamma correction)\n\
 		#endif\n\
 		lightMapColor *= LightMapColor;\n\
 	#else\n\
-		#define lightMapColor highp vec3(0.0, 0.0, 0.0)\n\
+		#define lightMapColor mediump vec3(0.0, 0.0, 0.0)\n\
 	#endif\n\
 \n\
 	// Get the emissive color\n\
 	#ifdef FS_EMISSIVEMAP\n\
 		lowp vec3 emissiveColor = texture2D(EmissiveMap, TexCoord0).rgb;\n\
 		#ifdef FS_GAMMACORRECTION\n\
-			emissiveColor = pow(emissiveColor, highp vec3(2.2, 2.2, 2.2));	// Perform sRGB to linear space conversion (gamma correction)\n\
+			emissiveColor = pow(emissiveColor, mediump vec3(2.2, 2.2, 2.2));	// Perform sRGB to linear space conversion (gamma correction)\n\
 		#endif\n\
 		emissiveColor *= EmissiveMapColor;\n\
 	#else\n\
-		#define emissiveColor highp vec3(0.0, 0.0, 0.0)\n\
+		#define emissiveColor mediump vec3(0.0, 0.0, 0.0)\n\
 	#endif\n\
 \n\
 	// Get the diffuse color\n\
@@ -368,7 +368,7 @@ void main()\n\
 				discard; // Throw the fragment away and don't draw it!\n\
 		#endif\n\
 		#ifdef FS_GAMMACORRECTION\n\
-			diffuseMapTexel.rgb = pow(diffuseMapTexel.rgb, highp vec3(2.2, 2.2, 2.2));	// Perform sRGB to linear space conversion (gamma correction)\n\
+			diffuseMapTexel.rgb = pow(diffuseMapTexel.rgb, mediump vec3(2.2, 2.2, 2.2));	// Perform sRGB to linear space conversion (gamma correction)\n\
 		#endif\n\
 		lowp vec3 diffuseColor = diffuseMapTexel.rgb*DiffuseColor.rgb;\n\
 	#else\n\
@@ -418,10 +418,10 @@ void main()\n\
 			lowp vec3 normal = normalize(VertexNormalVS);\n\
 		#endif\n\
 		if (normal.x != normal.x || normal.y != normal.y || normal.z != normal.z)	// IEEE standard: NaN != NaN - I don't use isnan so I can use lower shader versions\n\
-			normal = highp vec3(0.0, 0.0, 1.0); // I had situations with invalid normal vectors...\n\
+			normal = mediump vec3(0.0, 0.0, 1.0); // I had situations with invalid normal vectors...\n\
 	#else\n\
 		// Define a dummy normal so, when using reflections, we get at least some 'kind of reflection' effect\n\
-		#define normal highp vec3(0.0, 0.0, 1.0)\n\
+		#define normal mediump vec3(0.0, 0.0, 1.0)\n\
 	#endif\n\
 \n\
 	// Apply reflection - modifies the diffuse color\n\
@@ -438,16 +438,16 @@ void main()\n\
 		#endif\n\
 \n\
 		// Reflection color\n\
-		lowp vec3 reflectionColor = highp vec3(1.0, 1.0, 1.0);\n\
+		lowp vec3 reflectionColor = mediump vec3(1.0, 1.0, 1.0);\n\
 		#ifdef FS_2DREFLECTIONMAP\n\
 			// Spherical environment mapping\n\
-			highp vec3  r = ViewSpaceToWorldSpace*normalize(reflect(VertexPositionVS, normal));\n\
-			highp float m = 2.0*sqrt(r.x*r.x + r.y*r.y + (r.z + 1.0)*(r.z + 1.0));\n\
+			mediump vec3  r = ViewSpaceToWorldSpace*normalize(reflect(VertexPositionVS, normal));\n\
+			mediump float m = 2.0*sqrt(r.x*r.x + r.y*r.y + (r.z + 1.0)*(r.z + 1.0));\n\
 			#define FLT_MIN 1.175494351e-38 // Minimum positive value\n\
 			if (m < FLT_MIN)\n\
 				m = FLT_MIN;\n\
 			#undef FLT_MIN\n\
-			reflectionColor = texture2D(ReflectionMap, highp vec2(r.x/m + 0.5, 1.0 - (r.y/m + 0.5))).rgb;\n\
+			reflectionColor = texture2D(ReflectionMap, mediump vec2(r.x/m + 0.5, 1.0 - (r.y/m + 0.5))).rgb;\n\
 		#elif defined(FS_CUBEREFLECTIONMAP)\n\
 			// Cubic environment mapping\n\
 			// There's no need to normalize the reflection vector when using cube maps\n\
@@ -455,7 +455,7 @@ void main()\n\
 		#endif\n\
 		// Perform sRGB to linear space conversion (gamma correction)\n\
 		#ifdef FS_GAMMACORRECTION\n\
-			reflectionColor = pow(reflectionColor, highp vec3(2.2, 2.2, 2.2));\n\
+			reflectionColor = pow(reflectionColor, mediump vec3(2.2, 2.2, 2.2));\n\
 		#endif\n\
 		// Apply reflection color\n\
 		reflectionColor *= ReflectionColor;\n\
@@ -477,7 +477,7 @@ void main()\n\
 				#define specularExponent SpecularExponent\n\
 			#endif\n\
 		#else\n\
-			#define specularColor highp vec3(0.0, 0.0, 0.0)\n\
+			#define specularColor mediump vec3(0.0, 0.0, 0.0)\n\
 			#define specularExponent 0.0\n\
 		#endif\n\
 \n\
@@ -501,7 +501,7 @@ void main()\n\
 	// DOF\n\
 	#ifdef FS_DOF\n\
 		// Compute depth blur\n\
-		highp float f = -VertexPositionVS.z;\n\
+		mediump float f = -VertexPositionVS.z;\n\
 \n\
 		if (f < DOFParams.y) {\n\
 			// Scale depth value between near blur distance and focal distance to [-1, 0] range\n\
@@ -519,6 +519,6 @@ void main()\n\
 	#endif\n\
 \n\
 	// Calculate the fragment color\n\
-	gl_FragColor = highp vec4(ambientColor*diffuseColor + lightMapColor*diffuseColor + emissiveColor + lightingColor, alphaValue);\n\
+	gl_FragColor = mediump vec4(ambientColor*diffuseColor + lightMapColor*diffuseColor + emissiveColor + lightingColor, alphaValue);\n\
 \n\
 }";
